Separate form components into different files for reusability
